Fixing up syntax, fixing stupid valigrind error
Because this is supposed to be horrifying anyway, I decided to fix a
problem that wasn't.

Valgrind was finding an invalid read size of 8 after the realloc'd block
in resize for the second split loop. This is because we have an
exact-size array for our elements, so when we get to the final iteration
loop, before we check i < count, we try via object_vat to reach past.
So, I added a nice little MIN() call to degrade performance for no
benefit. Enjoy!